Modified the tests, fixed the auth error
