+++ RoomAid/RoomAid.ServiceLayer/App.config
@@ -87,10 +87,10 @@
     <add key="sqlConnectionSystem" value="Data Source=LAPTOP-BLNANITI;Initial Catalog=testsystemdb;User ID=dbtester;Password=Adjjmp491" />
     <add key="sqlConnectionMapping" value="Data Source=LAPTOP-BLNANITI;Initial Catalog=testmappingdb;User ID=dbtester;Password=Adjjmp491" />
     <add key ="updateCmd" value ="UPDATE dbo.Users SET FirstName = @fName, LastName = @lName, DateOfBirth = @dob, Gender = @gender, AccountStatus = @status WHERE UserEmail = @email"/>
-    <add key ="queryCreateAccount" value ="INSERT INTO dbo.Accounts (UserEmail, HashedPassword, Salt) VALUES (@email, @hashedPw, @salt)"/>
+    <add key ="queryCreateAccount" value ="INSERT INTO dbo.Accounts (UserEmail, HashPassword, Salt) VALUES (@email, @hashedPw, @salt)"/>
     <add key ="queryCreateMapping" value ="INSERT INTO dbo.Mapping (UserEmail) VALUES (@email)"/>
     <add key ="querySelectMapping" value ="SELECT SysID FROM dbo.Mapping WHERE UserEmail = @userEmail"/>
-    <add key ="queryCreateUser" value ="INSERT INTO dbo.Accounts (UserEmail, HashedPassword, Salt) VALUES (@email, @hashedPw, @salt)"/>
+    <add key ="queryCreateUser" value ="INSERT INTO dbo.Accounts (UserEmail, HashPassword, Salt) VALUES (@email, @hashedPw, @salt)"/>
     <add key="successMessage" value ="All targeted rows have been updated successfully!"/>
     <add key="failureMessage" value ="Not all targeted rows have been updated, please review which rows were attempted at updates"/>
 </appSettings>
