+++ RoomAid/RoomAid.UserManagement.Tests/RoomAid.UserManagement.Tests.csproj
@@ -46,10 +46,12 @@
       <HintPath>..\packages\MSTest.TestFramework.1.3.2\lib\net45\Microsoft.VisualStudio.TestPlatform.TestFramework.Extensions.dll</HintPath>
     </Reference>
     <Reference Include="System" />
+    <Reference Include="System.Configuration" />
     <Reference Include="System.Core" />
     <Reference Include="System.Data" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CreateAccountTests.cs" />
     <Compile Include="UpdateAccountSqlDAOTest.cs" />
     <Compile Include="UpdateAccountSqlServiceTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
