+++ RoomAid/RoomAid.sln
@@ -17,8 +17,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RoomAid.ManagerLayer", "Roo
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RoomAid.DataAccessLayer", "RoomAid.DataAccessLayer\RoomAid.DataAccessLayer.csproj", "{5E8E8769-AF6B-49C7-A61C-AA85B747EF93}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RoomAid.CreateAccount.Tests", "RoomAid.CreateAccount.Tests\RoomAid.CreateAccount.Tests.csproj", "{0DF67A7E-5571-49E7-8AF9-80348223E049}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RoomAid.ServiceLayer", "RoomAid.ServiceLayer\RoomAid.ServiceLayer.csproj", "{3FE754D4-2138-454A-9DE6-C7E3301E8D6C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RoomAid.DataMasking.Tests", "RoomAid.DataMasking.Tests\RoomAid.DataMasking.Tests.csproj", "{7854A54E-A169-4F0A-B9E3-55220EFA21D7}"
@@ -65,10 +63,6 @@ Global
 		{5E8E8769-AF6B-49C7-A61C-AA85B747EF93}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5E8E8769-AF6B-49C7-A61C-AA85B747EF93}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5E8E8769-AF6B-49C7-A61C-AA85B747EF93}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0DF67A7E-5571-49E7-8AF9-80348223E049}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0DF67A7E-5571-49E7-8AF9-80348223E049}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0DF67A7E-5571-49E7-8AF9-80348223E049}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0DF67A7E-5571-49E7-8AF9-80348223E049}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3FE754D4-2138-454A-9DE6-C7E3301E8D6C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3FE754D4-2138-454A-9DE6-C7E3301E8D6C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3FE754D4-2138-454A-9DE6-C7E3301E8D6C}.Release|Any CPU.ActiveCfg = Release|Any CPU
